refactor: extract EmailMessage and IncomingMessage into message module
Move EmailMessage and IncomingMessage from storage.rs into a dedicated
message.rs module with module-level documentation linking to RFC 5322.
storage.rs now only contains storage engine traits and implementations.
