+++ crates/utils/src/lib.rs
@@ -5,6 +5,7 @@ pub mod file;
 pub mod handler;
 pub mod handlers;
 pub mod imap;
+pub mod message;
 pub mod metadata;
 pub mod mime;
 pub mod router;
@@ -20,6 +21,7 @@ pub use file::*;
 pub use handler::*;
 pub use handlers::*;
 pub use imap::*;
+pub use message::*;
 pub use metadata::*;
 pub use mime::*;
 pub use router::*;
